Move regression test to regression directory
